Move into model/ folder
