+++ src/cell_annotator/__init__.py
@@ -1,11 +1,11 @@
 from importlib.metadata import version
 
-from ._api_keys import APIKeyManager
 from ._logging import logger
-from .base_annotator import BaseAnnotator
-from .cell_annotator import CellAnnotator
 from .check import check_deps
-from .sample_annotator import SampleAnnotator
+from .model._api_keys import APIKeyManager
+from .model.base_annotator import BaseAnnotator
+from .model.cell_annotator import CellAnnotator
+from .model.sample_annotator import SampleAnnotator
 
 __all__ = ["CellAnnotator", "SampleAnnotator", "BaseAnnotator", "logger", "check_deps", "APIKeyManager"]
 __version__ = version("cell-annotator")
