+++ tests/model/test_cell_annotator.py
